Added assigned connections attribute in connspec to store uint64_t connection counts to assign avoiding conversion issues

+++ src/connect_rules.cu
@@ -687,7 +687,7 @@ int NESTGPU::manual_assign_connections(
   inode_t* targets,
   float* weights,
   float* delays,
-  int count,
+  uint64_t count,
   bool remote_source_flag,
   SynSpec& syn_spec
 )
